+++ src/ts/QuerySelectors.ts
@@ -11,27 +11,22 @@
 
 export enum QuerySelectors {
     // Post
-    postWrapper = '.M9sTE, .NI8nC, article._aalr',
+    postWrapper = '.M9sTE, .NI8nC, article._aalr, article._aa6a, article._aatb',
     postBookmark = '.wmtNn, ._aamz > div',
-    postLink = '.c-Yi7, ._aat8 a',
-    postSliderIndicator = '.JSZAJ, .ijCUd',
-    postSliderActive = '.XCodT',
+    postSliderBubble = '.JSZAJ, .ijCUd, ._acnb',
+    postAccountName = "._aacl > ._aap6 >.oajrlxb2",
+    postSliderBubbleActive = '.XCodT, ._acnb._acnf',
+    sliderItem = "._acaz",
+    postContentWrapper = "._aatk._aatl, ._aatk._aatn, ._aato._ab1k._ab1l",
 
     // Story
-    storyAccountName = "._ac0q > .oajrlxb2",
     storyImage = '.y-yJ5, img._aa63',
     storyCloseButton = '.K_10X, ._g3zU, .aUIsh, ._ac0g>._abl-',
 
     // Account, Hover, Explore, Reels
-    imagePreview = '._bz0w, .pKKVh, .Tjpra > a',
     accountImage = '._aarf._aarg  ._aa8j',
     accountName = '._7UhW9.fKFbl.yUEEX.KV-D4.fDxYl, ._aacl._aacs._aact._aacx._aada',
-    topRightIconRow = '._47KiJ, .r9-Os, .J5g42',
 
     // Account Image
     accountImageWrapper = '.RR-M-, .M-jxE, ._aarf._aarg',
-
-    // Bulk download
-    loadingButton = '.By4nA',
-    stopDownload = '._0mzm-.sqdOP.yWX7d',
 }